feat : 불필요 파일, 메소드 제거, KeycloakService 이름 변경
